feat: TLS configs & custom client timeout
